Move semaphore to utils
