+++ modules/nlu/src/backend/application/tests/utils/fake-training-repo.u.test.ts
@@ -1,67 +1,6 @@
 import { TrainingState, TrainingId, TrainingSession } from '../../typings'
 import { ITrainingRepository, ITransactionContext } from '../../training-repo'
-
-type SemaphoreResolver = () => void
-
-interface SemaphoreQueueEntry {
-  resolve: (nextCycle: [number, SemaphoreResolver]) => void
-  reject: (err: Error) => void
-}
-class Semaphore {
-  private _value: number
-  private _queue: SemaphoreQueueEntry[] = []
-  private _resolver: SemaphoreResolver | undefined
-
-  constructor(private _n: number) {
-    if (_n <= 0) {
-      throw new Error('Semaphore must be initialized with a positive value')
-    }
-    this._value = _n
-  }
-
-  get locked() {
-    return this._value <= 0
-  }
-
-  async exclusively(cb: (val: number) => Promise<any>) {
-    const [value, unlock] = await this._lock()
-    try {
-      return cb(value)
-    } finally {
-      unlock()
-    }
-  }
-
-  private _lock(): Promise<[number, SemaphoreResolver]> {
-    const lockPromise = new Promise<[number, SemaphoreResolver]>((resolve, reject) => {
-      this._queue.push({ resolve, reject })
-    })
-
-    if (!this.locked) this._cycle()
-    return lockPromise
-  }
-
-  private _cycle() {
-    const promise = this._queue.shift()
-
-    if (!promise) {
-      return
-    }
-
-    let unlocked = false
-
-    this._resolver = () => {
-      if (unlocked) {
-        return
-      }
-      unlocked = true
-      this._value++
-      this._cycle()
-    }
-    promise.resolve([this._value--, this._resolver])
-  }
-}
-
+import { Semaphore } from './utils.u.test'
 class FakeTransactionContext implements ITransactionContext {
   public transaction = null
   private _trainings: { [key: string]: TrainingState & { modifiedOn: Date } } = {}
